pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sf-school-navigator-pipeline"
version = "0.1.0"
description = "Data pipeline for SF School Navigator"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "supabase>=2.0",
    "pydantic>=2.0",
    "click>=8.1",
    "python-slugify>=8.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
pipeline = "pipeline.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
